refactor: isolate runtime from compiler
