+++ src/public/result.ts
@@ -1,0 +1,8 @@
+export type { Result, Ok, Err, QueryMeta } from '../runtime/result.js';
+export {
+  ResultStatus,
+  ok,
+  err,
+  isOk,
+  isErr,
+} from '../runtime/result.js';
