+++ tests/contracts/public-api.test-d.ts
@@ -1,11 +1,33 @@
 import type {
+  DialectExecutor,
+  Err,
+  Executor,
+  ExecutorResult,
+  FunctionReturnTypes,
+  InferParams,
+  InferResult,
+  InferResultStrict,
+  InferRow,
+  InferRowStrict,
+  Ok,
   Params,
+  ParseSelect,
+  ParseStatement,
+  ParsedStatement,
+  PlaceholderStyle,
   Query,
+  QueryError,
+  QueryMeta,
   QueryTypeError,
+  Result,
   Row,
+  Schema,
+  SchemaLike,
+  Source,
   StrictQuery,
   StrictRow,
   TypedDb,
+  TypedDbOptions,
 } from '../../src/index.js';
 
 type Equal<A, B> =
