+++ tests/contracts/runtime-contract.test.ts
@@ -6,12 +6,24 @@ import {
   isErr,
   isOk,
 } from '../../src/index.js';
+import { createRuntimeDb } from '../../src/runtime/db.js';
 
 interface DB {
   users: { id: number; name: string };
 }
 
 describe('public runtime contract', () => {
+  it('runtime db accepts plain strings without compiler types', async () => {
+    const db = createRuntimeDb(async (sql, params) => {
+      expect(sql).toBe('select 1');
+      expect(params).toEqual([7]);
+      return [{ value: 1 }];
+    });
+
+    const result = await db.query('select 1', [7]);
+    expect(isOk(result)).toBe(true);
+  });
+
   it('wraps rows in an OK Result', async () => {
     const db = createTypedDb<DB>(async () => [{ id: 1, name: 'Ada' }]);
     const result = await db.query('select id, name from users');
